+++ src/Controller/Api/UsersApiController.php
@@ -3,12 +3,14 @@
 namespace Kuusamo\Vle\Controller\Api;
 
 use Kuusamo\Vle\Entity\User;
+use Kuusamo\Vle\Entity\UserCourse;
 use Kuusamo\Vle\Validation\UserValidator;
 use Kuusamo\Vle\Validation\ValidationException;
 
 use Doctrine\DBAL\Exception\UniqueConstraintViolationException;
 use Psr\Http\Message\ServerRequestInterface as Request;
 use Psr\Http\Message\ResponseInterface as Response;
+use Slim\Exception\HttpNotFoundException;
 
 class UsersApiController extends ApiController
 {
@@ -38,5 +40,39 @@ class UsersApiController extends ApiController
         }
 
         return $response->withJson($user);
+    }
+
+    public function enrol(Request $request, Response $response, array $args = [])
+    {
+        if ($errorMsg = $this->verifyRequest($request)) {
+            return $this->badRequest($response, $errorMsg);
+        }
+
+        $user = $this->ci->get('db')->find('Kuusamo\Vle\Entity\User', $args['id']);
+
+        if ($user === null) {
+            throw new HttpNotFoundException($request, $response);
+        }
+
+        $json = $this->getJson($request);
+
+        $course = $this->ci->get('db')->find('Kuusamo\Vle\Entity\Course', $json['id']);
+
+        if ($course === null) {
+            return $this->badRequest('Course not found');
+        }
+
+        $userCourse = new UserCourse;
+        $userCourse->setUser($user);
+        $userCourse->setCourse($course);
+
+        try {
+            $this->ci->get('db')->persist($userCourse);
+            $this->ci->get('db')->flush();
+        } catch (UniqueConstraintViolationException $e) {
+            return $this->badRequest($response, 'User is already enrolled on this course');
+        }
+
+        return $response->withJson([]);
     }
 }