+++ changelog.md
@@ -9,7 +9,7 @@ Changelog
 * Graded assessments
 * Manual marking for students
 * API added with test end-point
-* User create API
+* User create and enrol API
 * Handle JSON requests in exception controller
 * Handle duplicate email address errors
 * Handle duplicate course URI errors